dive-log-service/requirements.txt
flask==2.2.3
mysql-connector-python==8.0.32
python-dotenv==1.0.0
flask-cors==3.0.10
